+++ benchmarks/source-datocms/gatsby-node.js
@@ -3,8 +3,8 @@ const kebabCase = require(`lodash.kebabcase`)
 exports.onCreateNode = ({ actions, node }) => {
   const { createNodeField } = actions
 
-  if (node.internal.type === 'node__article') {
-    createNodeField({ node, name: "slug", value: kebabCase(node.title) })
+  if (node.internal.type === `node__article`) {
+    createNodeField({ node, name: `slug`, value: kebabCase(node.title) })
   }
 }
 
@@ -16,7 +16,7 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
       articles: allDatoCmsArticle {
         nodes {
           id
-          path
+          slug
         }
       }
     }
@@ -28,11 +28,11 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
 
   result.data.articles.nodes.map(article => {
     createPage({
-      path: article.path,
+      path: article.slug,
       component: require.resolve(`./src/templates/article.js`),
       context: {
         id: article.id,
-      }
+      },
     })
   })
 }