+++ packages/gatsby-image/src/index.js
@@ -262,14 +262,12 @@ const noscriptImg = props => {
 // Earlier versions of gatsby-image during the 2.x cycle did not wrap
 // the `Img` component in a `picture` element. This maintains compatibility
 // until a breaking change can be introduced in the next major release
-const Placeholder = ({
-  src,
-  imageVariants,
-  generateSources,
-  spreadProps,
-  ariaHidden,
-}) => {
-  const baseImage = <Img src={src} {...spreadProps} ariaHidden={ariaHidden} />
+const Placeholder = React.forwardRef((props, ref) => {
+  const { src, imageVariants, generateSources, spreadProps, ariaHidden } = props
+
+  const baseImage = (
+    <Img ref={ref} src={src} {...spreadProps} ariaHidden={ariaHidden} />
+  )
 
   return imageVariants.length > 1 ? (
     <picture>
@@ -279,7 +277,7 @@ const Placeholder = ({
   ) : (
     baseImage
   )
-}
+})
 
 const Img = React.forwardRef((props, ref) => {
   const {
@@ -316,7 +314,7 @@ const Img = React.forwardRef((props, ref) => {
         height: `100%`,
         objectFit: `cover`,
         objectPosition: `center`,
-        ...style,
+        ...style
       }}
     />
   )
@@ -325,7 +323,7 @@ const Img = React.forwardRef((props, ref) => {
 Img.propTypes = {
   style: PropTypes.object,
   onError: PropTypes.func,
-  onLoad: PropTypes.func,
+  onLoad: PropTypes.func
 }
 
 class Image extends React.Component {
@@ -353,10 +351,11 @@ class Image extends React.Component {
       isVisible,
       imgLoaded: false,
       imgCached: false,
-      fadeIn: !this.seenBefore && props.fadeIn,
+      fadeIn: !this.seenBefore && props.fadeIn
     }
 
     this.imageRef = React.createRef()
+    this.placeholderRef = props.placeholderRef || React.createRef()
     this.handleImageLoaded = this.handleImageLoaded.bind(this)
     this.handleRef = this.handleRef.bind(this)
   }
@@ -404,7 +403,7 @@ class Image extends React.Component {
             // TODO fix imgCached behaviour as it's now false when it's lazyloaded
             imgCached: !!(
               this.imageRef.current && this.imageRef.current.currentSrc
-            ),
+            )
           })
         })
       })
@@ -437,7 +436,7 @@ class Image extends React.Component {
       Tag,
       itemProp,
       loading,
-      draggable,
+      draggable
     } = convertProps(this.props)
 
     const shouldReveal = this.state.fadeIn === false || this.state.imgLoaded
@@ -446,21 +445,21 @@ class Image extends React.Component {
     const imageStyle = {
       opacity: shouldReveal ? 1 : 0,
       transition: shouldFadeIn ? `opacity ${durationFadeIn}ms` : `none`,
-      ...imgStyle,
+      ...imgStyle
     }
 
     const bgColor =
       typeof backgroundColor === `boolean` ? `lightgray` : backgroundColor
 
     const delayHideStyle = {
-      transitionDelay: `${durationFadeIn}ms`,
+      transitionDelay: `${durationFadeIn}ms`
     }
 
     const imagePlaceholderStyle = {
       opacity: this.state.imgLoaded ? 0 : 1,
       ...(shouldFadeIn && delayHideStyle),
       ...imgStyle,
-      ...placeholderStyle,
+      ...placeholderStyle
     }
 
     const placeholderImageProps = {
@@ -468,7 +467,7 @@ class Image extends React.Component {
       alt: !this.state.isVisible ? alt : ``,
       style: imagePlaceholderStyle,
       className: placeholderClassName,
-      itemProp,
+      itemProp
     }
 
     if (fluid) {
@@ -481,7 +480,7 @@ class Image extends React.Component {
           style={{
             position: `relative`,
             overflow: `hidden`,
-            ...style,
+            ...style
           }}
           ref={this.handleRef}
           key={`fluid-${JSON.stringify(image.srcSet)}`}
@@ -491,7 +490,7 @@ class Image extends React.Component {
             aria-hidden
             style={{
               width: `100%`,
-              paddingBottom: `${100 / image.aspectRatio}%`,
+              paddingBottom: `${100 / image.aspectRatio}%`
             }}
           />
 
@@ -508,7 +507,7 @@ class Image extends React.Component {
                 opacity: !this.state.imgLoaded ? 1 : 0,
                 right: 0,
                 left: 0,
-                ...(shouldFadeIn && delayHideStyle),
+                ...(shouldFadeIn && delayHideStyle)
               }}
             />
           )}
@@ -517,6 +516,7 @@ class Image extends React.Component {
           {image.base64 && (
             <Placeholder
               ariaHidden
+              ref={this.placeholderRef}
               src={image.base64}
               spreadProps={placeholderImageProps}
               imageVariants={imageVariants}
@@ -528,6 +528,7 @@ class Image extends React.Component {
           {image.tracedSVG && (
             <Placeholder
               ariaHidden
+              ref={this.placeholderRef}
               src={image.tracedSVG}
               spreadProps={placeholderImageProps}
               imageVariants={imageVariants}
@@ -566,8 +567,8 @@ class Image extends React.Component {
                   title,
                   loading,
                   ...image,
-                  imageVariants,
-                }),
+                  imageVariants
+                })
               }}
             />
           )}
@@ -585,7 +586,7 @@ class Image extends React.Component {
         display: `inline-block`,
         width: image.width,
         height: image.height,
-        ...style,
+        ...style
       }
 
       if (style.display === `inherit`) {
@@ -609,7 +610,7 @@ class Image extends React.Component {
                 width: image.width,
                 opacity: !this.state.imgLoaded ? 1 : 0,
                 height: image.height,
-                ...(shouldFadeIn && delayHideStyle),
+                ...(shouldFadeIn && delayHideStyle)
               }}
             />
           )}
@@ -618,6 +619,7 @@ class Image extends React.Component {
           {image.base64 && (
             <Placeholder
               ariaHidden
+              ref={this.placeholderRef}
               src={image.base64}
               spreadProps={placeholderImageProps}
               imageVariants={imageVariants}
@@ -629,6 +631,7 @@ class Image extends React.Component {
           {image.tracedSVG && (
             <Placeholder
               ariaHidden
+              ref={this.placeholderRef}
               src={image.tracedSVG}
               spreadProps={placeholderImageProps}
               imageVariants={imageVariants}
@@ -669,8 +672,8 @@ class Image extends React.Component {
                   title,
                   loading,
                   ...image,
-                  imageVariants,
-                }),
+                  imageVariants
+                })
               }}
             />
           )}
@@ -689,7 +692,7 @@ Image.defaultProps = {
   Tag: `div`,
   // We set it to `lazy` by default because it's best to default to a performant
   // setting and let the user "opt out" to `eager`
-  loading: `lazy`,
+  loading: `lazy`
 }
 
 const fixedObject = PropTypes.shape({
@@ -701,7 +704,7 @@ const fixedObject = PropTypes.shape({
   tracedSVG: PropTypes.string,
   srcWebp: PropTypes.string,
   srcSetWebp: PropTypes.string,
-  media: PropTypes.string,
+  media: PropTypes.string
 })
 
 const fluidObject = PropTypes.shape({
@@ -713,7 +716,7 @@ const fluidObject = PropTypes.shape({
   tracedSVG: PropTypes.string,
   srcWebp: PropTypes.string,
   srcSetWebp: PropTypes.string,
-  media: PropTypes.string,
+  media: PropTypes.string
 })
 
 // If you modify these propTypes, please don't forget to update following files as well:
@@ -743,7 +746,7 @@ Image.propTypes = {
   Tag: PropTypes.string,
   itemProp: PropTypes.string,
   loading: PropTypes.oneOf([`auto`, `lazy`, `eager`]),
-  draggable: PropTypes.bool,
+  draggable: PropTypes.bool
 }
 
 export default Image