+++ packages/gatsby-transformer-remark/src/extend-node-type.js
@@ -18,7 +18,7 @@ const prune = require(`underscore.string/prune`)
 const {
   getConcatenatedValue,
   cloneTreeUntil,
-  findLastTextNode,
+  findLastTextNode
 } = require(`./hast-processing`)
 const codeHandler = require(`./code-handler`)
 const { timeToRead } = require(`./utils/time-to-read`)
@@ -71,7 +71,7 @@ const SpaceMarkdownNodeTypesSet = new Set([
   `paragraph`,
   `heading`,
   `tableCell`,
-  `break`,
+  `break`
 ])
 
 const headingLevels = [...Array(6).keys()].reduce((acc, i) => {
@@ -110,15 +110,15 @@ module.exports = (
       pedantic = true,
       tableOfContents = {
         heading: null,
-        maxDepth: 6,
-      },
+        maxDepth: 6
+      }
     } = pluginOptions
     const tocOptions = tableOfContents
     const remarkOptions = {
       commonmark,
       footnotes,
       gfm,
-      pedantic,
+      pedantic
     }
     if (_.isArray(blocks)) {
       remarkOptions.blocks = blocks
@@ -183,9 +183,9 @@ module.exports = (
               getCache,
               compiler: {
                 parseString: remark.parse.bind(remark),
-                generateHTML: getHTML,
+                generateHTML: getHTML
               },
-              ...rest,
+              ...rest
             },
             plugin.pluginOptions
           )
@@ -234,9 +234,9 @@ module.exports = (
               getCache,
               compiler: {
                 parseString: remark.parse.bind(remark),
-                generateHTML: getHTML,
+                generateHTML: getHTML
               },
-              ...rest,
+              ...rest
             },
             plugin.pluginOptions
           )
@@ -257,7 +257,7 @@ module.exports = (
         const headings = select(ast, `heading`).map(heading => {
           return {
             value: mdastToString(heading),
-            depth: heading.depth,
+            depth: heading.depth
           }
         })
 
@@ -295,7 +295,7 @@ module.exports = (
               node.url = [
                 basePath,
                 _.get(markdownNode, appliedTocOptions.pathToSlugField),
-                node.url,
+                node.url
               ]
                 .join(`/`)
                 .replace(/\/\//g, `/`)
@@ -313,7 +313,7 @@ module.exports = (
           }
 
           toc = hastToHTML(toHAST(tocAst.map, { allowDangerousHTML: true }), {
-            allowDangerousHTML: true,
+            allowDangerousHTML: true
           })
         } else {
           toc = ``
@@ -331,7 +331,7 @@ module.exports = (
         const ast = await getAST(markdownNode)
         const htmlAst = toHAST(ast, {
           allowDangerousHTML: true,
-          handlers: { code: codeHandler },
+          handlers: { code: codeHandler }
         })
 
         // Save new HTML AST to cache and return
@@ -350,7 +350,7 @@ module.exports = (
         const ast = await getHTMLAst(markdownNode)
         // Save new HTML to cache and return
         const html = hastToHTML(ast, {
-          allowDangerousHTML: true,
+          allowDangerousHTML: true
         })
 
         if (shouldCache) {
@@ -403,7 +403,7 @@ module.exports = (
       } else {
         lastTextNode.value = _.truncate(lastTextNode.value, {
           length: pruneLength,
-          omission: `…`,
+          omission: `…`
         })
       }
       return excerptAST
@@ -419,10 +419,10 @@ module.exports = (
       const excerptAST = await getExcerptAst(fullAST, markdownNode, {
         pruneLength,
         truncate,
-        excerptSeparator,
+        excerptSeparator
       })
       const html = hastToHTML(excerptAST, {
-        allowDangerousHTML: true,
+        allowDangerousHTML: true
       })
       return html
     }
@@ -441,7 +441,7 @@ module.exports = (
       const excerptAST = await getExcerptAst(ast, markdownNode, {
         pruneLength,
         truncate,
-        excerptSeparator,
+        excerptSeparator
       })
       var excerptMarkdown = unified()
         .use(stringify)
@@ -485,7 +485,7 @@ module.exports = (
         }
         return _.truncate(excerptText, {
           length: pruneLength,
-          omission: `…`,
+          omission: `…`
         })
       })
       return text
@@ -523,7 +523,7 @@ module.exports = (
         type: `String`,
         resolve(markdownNode) {
           return getHTML(markdownNode)
-        },
+        }
       },
       htmlAst: {
         type: `JSON`,
@@ -532,44 +532,44 @@ module.exports = (
             const strippedAst = stripPosition(_.clone(ast), true)
             return hastReparseRaw(strippedAst)
           })
-        },
+        }
       },
       excerpt: {
         type: `String`,
         args: {
           pruneLength: {
             type: `Int`,
-            defaultValue: 140,
+            defaultValue: 140
           },
           truncate: {
             type: `Boolean`,
-            defaultValue: false,
+            defaultValue: false
           },
           format: {
             type: `MarkdownExcerptFormats`,
-            defaultValue: `PLAIN`,
-          },
+            defaultValue: `PLAIN`
+          }
         },
         resolve(markdownNode, { format, pruneLength, truncate }) {
           return getExcerpt(markdownNode, {
             format,
             pruneLength,
             truncate,
-            excerptSeparator: pluginOptions.excerpt_separator,
+            excerptSeparator: pluginOptions.excerpt_separator
           })
-        },
+        }
       },
       excerptAst: {
         type: `JSON`,
         args: {
           pruneLength: {
             type: `Int`,
-            defaultValue: 140,
+            defaultValue: 140
           },
           truncate: {
             type: `Boolean`,
-            defaultValue: false,
-          },
+            defaultValue: false
+          }
         },
         resolve(markdownNode, { pruneLength, truncate }) {
           return getHTMLAst(markdownNode)
@@ -577,19 +577,19 @@ module.exports = (
               getExcerptAst(fullAST, markdownNode, {
                 pruneLength,
                 truncate,
-                excerptSeparator: pluginOptions.excerpt_separator,
+                excerptSeparator: pluginOptions.excerpt_separator
               })
             )
             .then(ast => {
               const strippedAst = stripPosition(_.clone(ast), true)
               return hastReparseRaw(strippedAst)
             })
-        },
+        }
       },
       headings: {
         type: [`MarkdownHeading`],
         args: {
-          depth: `MarkdownHeadingLevels`,
+          depth: `MarkdownHeadingLevels`
         },
         resolve(markdownNode, { depth }) {
           return getHeadings(markdownNode).then(headings => {
@@ -599,13 +599,13 @@ module.exports = (
             }
             return headings
           })
-        },
+        }
       },
       timeToRead: {
         type: `Int`,
         resolve(markdownNode) {
           return getHTML(markdownNode).then(timeToRead)
-        },
+        }
       },
       tableOfContents: {
         type: `String`,
@@ -613,19 +613,19 @@ module.exports = (
           // TODO:(v3) set default value to false
           absolute: {
             type: `Boolean`,
-            defaultValue: true,
+            defaultValue: true
           },
           // TODO:(v3) set default value to empty string
           pathToSlugField: {
             type: `String`,
-            defaultValue: `fields.slug`,
+            defaultValue: `fields.slug`
           },
           maxDepth: `Int`,
-          heading: `String`,
+          heading: `String`
         },
         resolve(markdownNode, args) {
           return getTableOfContents(markdownNode, args)
-        },
+        }
       },
       // TODO add support for non-latin languages https://github.com/wooorm/remark/issues/251#issuecomment-296731071
       wordCount: {
@@ -647,7 +647,7 @@ module.exports = (
           return {
             paragraphs: counts.ParagraphNode,
             sentences: counts.SentenceNode,
-            words: counts.WordNode,
+            words: counts.WordNode
           }
 
           function count() {
@@ -659,8 +659,8 @@ module.exports = (
               }
             }
           }
-        },
-      },
+        }
+      }
     })
   })
 }