+++ packages/gatsby/src/query/graphql-runner.ts
@@ -1,3 +1,5 @@
+import crypto from "crypto"
+import v8 from "v8"
 import {
   parse,
   validate,
@@ -164,9 +166,17 @@ export default class GraphQLRunner {
       if (typeof statsQuery !== `string`) {
         statsQuery = statsQuery.body
       }
-      this.stats.uniqueOperations.add(`${statsQuery}${JSON.stringify(context)}`)
+      this.stats.uniqueOperations.add(
+        crypto
+          .createHash(`sha1`)
+          .update(statsQuery)
+          .update(v8.serialize(context))
+          .digest(`hex`)
+      )
 
-      this.stats.uniqueQueries.add(statsQuery)
+      this.stats.uniqueQueries.add(
+        crypto.createHash(`sha1`).update(statsQuery).digest(`hex`)
+      )
     }
 
     const document = this.parse(query)