+++ packages/gatsby/src/schema/resolvers.js
@@ -15,7 +15,7 @@ const findMany = typeName => (source, args, context, info) => {
       query: args,
       firstOnly: false,
       type: info.schema.getType(typeName),
-      stats: context.stats,
+      stats: context.stats
     },
     { path: context.path, connectionType: typeName }
   )
@@ -30,7 +30,7 @@ const findOne = typeName => (source, args, context, info) => {
       query: { filter: args },
       firstOnly: true,
       type: info.schema.getType(typeName),
-      stats: context.stats,
+      stats: context.stats
     },
     { path: context.path }
   )
@@ -74,14 +74,18 @@ const group = (source, args, context, info) => {
         acc[key] = (acc[key] || []).concat(node)
       })
     return acc
-  }, {})
+    // Note: using Object.create on purpose:
+    //   object key may be arbitrary string including reserved words (i.e. `constructor`)
+    //   see: https://github.com/gatsbyjs/gatsby/issues/22508
+  }, Object.create(null))
+
   return Object.keys(groupedResults)
     .sort()
     .reduce((acc, fieldValue) => {
       acc.push({
         ...paginate(groupedResults[fieldValue], args),
         field,
-        fieldValue,
+        fieldValue
       })
       return acc
     }, [])
@@ -110,7 +114,7 @@ const paginate = (results = [], { skip = 0, limit }) => {
       return {
         node: item,
         next: arr[i + 1],
-        previous: arr[i - 1],
+        previous: arr[i - 1]
       }
     }),
     nodes: items,
@@ -120,8 +124,8 @@ const paginate = (results = [], { skip = 0, limit }) => {
       hasNextPage,
       itemCount: items.length,
       pageCount,
-      perPage: limit,
-    },
+      perPage: limit
+    }
   }
 }
 
@@ -135,7 +139,7 @@ const link = (options = {}, fieldConfig) => async (
   const fieldValue = await resolver(source, args, context, {
     ...info,
     from: options.from || info.from,
-    fromNode: options.from ? options.fromNode : info.fromNode,
+    fromNode: options.from ? options.fromNode : info.fromNode
   })
 
   if (fieldValue == null) return null
@@ -172,7 +176,7 @@ const link = (options = {}, fieldConfig) => async (
   const operator = Array.isArray(fieldValue) ? oneOf : equals
   args.filter = options.by.split(`.`).reduceRight((acc, key, i, { length }) => {
     return {
-      [key]: i === length - 1 ? operator(acc) : acc,
+      [key]: i === length - 1 ? operator(acc) : acc
     }
   }, fieldValue)
 
@@ -212,7 +216,7 @@ const fileByPath = (options = {}, fieldConfig) => async (
   const fieldValue = await resolver(source, args, context, {
     ...info,
     from: options.from || info.from,
-    fromNode: options.from ? options.fromNode : info.fromNode,
+    fromNode: options.from ? options.fromNode : info.fromNode
   })
 
   if (fieldValue == null) return null
@@ -283,7 +287,7 @@ const getFieldNodeByNameInSelectionSet = (selectionSet, fieldName, info) =>
             fragmentDef.selectionSet,
             fieldName,
             info
-          ),
+          )
         ]
       }
     } else if (selection.kind === Kind.INLINE_FRAGMENT) {
@@ -293,7 +297,7 @@ const getFieldNodeByNameInSelectionSet = (selectionSet, fieldName, info) =>
           selection.selectionSet,
           fieldName,
           info
-        ),
+        )
       ]
     } /* FIELD_NODE */ else {
       if (selection.name.value === fieldName) {
@@ -326,5 +330,5 @@ module.exports = {
   link,
   distinct,
   group,
-  paginate,
+  paginate
 }